rename missing classes
