+++ kotest-common/src/jvmTest/kotlin/io/kotest/equals/ReflectionIgnoringFieldsEqualityTest.kt
@@ -1,16 +1,15 @@
-package io.kotest.comparators
+package io.kotest.equals
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.data.forAll
 import io.kotest.data.headers
 import io.kotest.data.row
 import io.kotest.data.table
-import io.kotest.equals.Equality
 import io.kotest.matchers.booleans.shouldBeFalse
 import io.kotest.matchers.booleans.shouldBeTrue
 import io.kotest.matchers.shouldBe
 
-class ReflectionIgnoringFieldsComparatorTest : FunSpec({
+class ReflectionIgnoringFieldsEqualityTest : FunSpec({
    data class Foo(val a: String, val b: Int, val c: Boolean)
 
    data class Car(val name: String, val price: Int, private val modelNumber: Int)
